Azure_Technology_45_Minute_Demo_Content/demo-content/Real-Time-Remote-Patient-Monitoring-content/functions/requirements.txt
azure-functions
azure-identity>=1.15.0
requests>=2.31.0
